// schemas/integrations-schema.ts
import z from "zod";

export const formSchema = z.object({
    apiKey: z.string().default(""),
    apiUrl: z.string().default(""),
    region: z.string().default("australia"),
    teamId: z.string().default(""),
    authType: z.string().default("OAuth2.0"),
    credName: z.string().default(""),
    tokenUrl: z.string().default(""),
    clientId: z.string().default(""),
    clientSecret: z.string().default(""),
    scope: z.string().default(""),
    fallbackDisabled: z.boolean().default(false),
    sbBucketName: z.string().default(""),
    sbBucketRegion: z.string().default(""),
    sbPathPrefix: z.string().default(""),
    sbEndpoint: z.string().default(""),
    sbAccessKey: z.string().default(""),
    sbSecretKey: z.string().default(""),
    sbFallbackIndex: z.string().default(""),
    gcpRefName: z.string().default(""),
    gcpJson: z.string().default(""),
    gcpBucketName: z.string().default(""),
    gcpBucketRegion: z.string().default(""),
    gcpPathPrefix: z.string().default(""),
    gcpHmacAccessKey: z.string().default(""),
    gcpHmacSecret: z.string().default(""),
    gcpFallbackIndex: z.string().default(""),
    awsAccessKeyId: z.string().default(""),
    awsSecretAccessKey: z.string().default(""),
    s3BucketRegion: z.string().default(""),
    s3BucketName: z.string().default(""),
    s3PathPrefix: z.string().default(""),
    fallbackIndex: z.string().default(""),
})